+++ common/api/delays.ts
@@ -1,5 +1,5 @@
 import { FetchAlertDelaysByLineParams, type FetchAlertDelaysByLineOptions } from '../types/api';
-import type { LineDelays } from '../types/reliability';
+import type { LineDelays } from '../types/delays';
 import { apiFetch } from './utils/fetch';
 
 export const fetchLineDelaysByLine = async (
@@ -10,6 +10,6 @@ export const fetchLineDelaysByLine = async (
   return await apiFetch({
     path: '/api/linedelays',
     options,
-    errorMessage: 'Failed to fetch reliability metrics',
+    errorMessage: 'Failed to fetch delay metrics',
   });
 };
